@@ -5,7 +5,7 @@ author: Leonid Veremchuk<leonid.veremchuk@gmail.com>
 homepage: https://github.com/LeonidVeremchuk
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
